chore: rename TUI files to kebab case
