+++ src/tui/components/trace-item-preview.js
