+++ src/tui/event-decorators/apply-patch.js
