+++ src/tui/event-decorators/write-file.js
