+++ src/tui/hooks/use-selection.js
