+++ src/tui/hooks/use-stdout-dimensions.js
