+++ src/tui/hooks/use-terminal-mouse-mode.js
