+++ src/tui/views/trace-types-filter-view.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useInput, Text } from 'ink';
 import { MultiSelect } from '@inkjs/ui';
-import { useStdoutDimensions } from '../hooks/useStdoutDimensions.js';
-import { AppContainer } from '../components/AppContainer.js';
+import { useStdoutDimensions } from '../hooks/use-stdout-dimensions.js';
+import { AppContainer } from '../components/app-container.js';
 import { DEFAULT_FILTERS } from '../ui-utils/entry-utils.js';
 import { SHORTCUTS, buildFooter } from '../ui-utils/constants.js';
 import { isBackKey } from '../ui-utils/text-utils.js';
