+++ test/tui/views/project-selector/project-selector.test.js
@@ -5,7 +5,7 @@ import { render } from 'ink-testing-library';
 import fs from 'fs-extra';
 import path from 'path';
 import os from 'os';
-import { ProjectSelector } from '../../../../src/tui/views/ProjectSelector.js';
+import { ProjectSelector } from '../../../../src/tui/views/project-selector.js';
 
 describe('ProjectSelector ink-ui integration', () => {
   let tmpHome;
